src/proposal_generator/components/mockup_generator.py
"""Module for generating website mockups and design proposals."""

import logging
from typing import Dict, Any, List, Optional
from pathlib import Path

logger = logging.getLogger(__name__)

class MockupGenerator:
    """Generator for website mockups and design proposals."""

    def __init__(self):
        """Initialize the MockupGenerator."""
        self.logger = logger
        self.mockups_dir = Path(__file__).parent.parent.parent / 'mockups'

    def generate_mockup(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Generate a mockup based on input data.
        
        Args:
            data: Dictionary containing mockup requirements
            
        Returns:
            Dictionary containing the generated mockup and paths to mockup files
        """
        try:
            mockup_data = {
                'status': 'success',
                'mockup': {
                    'layout': self._generate_layout(data),
                    'design': self._generate_design(data),
                    'content': self._generate_content(data)
                },
                'files': self._get_mockup_files(data.get('page_type', 'homepage'))
            }
            return mockup_data
        except Exception as e:
            self.logger.error(f"Error generating mockup: {str(e)}")
            return {
                'status': 'error',
                'message': str(e)
            }

    def _get_mockup_files(self, page_type: str) -> Dict[str, str]:
        """Get mockup file paths for the specified page type."""
        return {
            'desktop': str(self.mockups_dir / f'{page_type}_mockup.png'),
            'mobile': str(self.mockups_dir / f'{page_type}_mobile_mockup.png')
        }

    def _generate_layout(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """Generate layout structure."""
        return {
            'type': 'layout',
            'sections': self._create_sections(data)
        }

    def _generate_design(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """Generate design elements."""
        return {
            'type': 'design',
            'elements': self._create_design_elements(data)
        }

    def _generate_content(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """Generate content placeholders."""
        return {
            'type': 'content',
            'elements': self._create_content_elements(data)
        }

    def _create_sections(self, data: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Create layout sections based on requirements."""
        return [
            {'name': 'header', 'type': 'navigation'},
            {'name': 'main', 'type': 'content'},
            {'name': 'footer', 'type': 'information'}
        ]

    def _create_design_elements(self, data: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Create design elements based on requirements."""
        return [
            {'type': 'color_scheme', 'values': ['#primary', '#secondary', '#accent']},
            {'type': 'typography', 'font_family': 'System Default'},
            {'type': 'spacing', 'unit': 'rem'}
        ]

    def _create_content_elements(self, data: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Create content elements based on requirements."""
        return [
            {'type': 'text', 'placeholder': 'Main heading'},
            {'type': 'image', 'placeholder': 'Hero image'},
            {'type': 'button', 'placeholder': 'Call to action'}
        ]

# Make sure the class is exported
__all__ = ['MockupGenerator']